requirements.txt
django==5.0.2
python-dotenv==1.0.1
django-allauth==0.61.1
python-decouple==3.8
whitenoise==6.6.0
google-generativeai==0.8.5 